refactor: update all paths of the interfaces
