+++ src/App/Pages/CoinMart/Components/Header/Header.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import Choose from "./Components/Choose";
 import Navigate from "./Components/Navigate";
 import Search from "./Components/Search";
-import { ICoinMartQuery } from "./Interfaces/ICoinMartQuery";
+import { ICoinMartQuery } from "./Interface/ICoinMartQuery";
 import styleHeader from "./styleHeader.module.scss";
 
 const Header: React.FC<ICoinMartQuery> = ({ coinmartquery }) => {
